feat(m68000): *toSR (supervisor) + RESET — privileged SR ops (189573 SST cases)
E6 of the 68000-exceptions slice. Adds the three privileged immediate-to-SR
logic ops and RESET over the proven micro-op framework.

- microop: new MicroOp::SrLogic{op: LogicOp{And,Or,Eor}, value} —
  sr = (sr <op> resolve(value)) & SR_IMPLEMENTED (0xA71F); the whole SR is
  rewritten, so And/Eor can clear S (switch supervisor->user) or T. Widen
  MicroOp::Internal.cycles to u16 (RESET idles the reset line 124 cycles).
- decode: ANDItoSR(0x027C)/ORItoSR(0x007C)/EORItoSR(0x0A7C) ->
  [Read(PcPlus(4),Program -> discard), Internal(8), SrLogic(op, ImmWord),
  Prefetch, Prefetch]. The discard read runs under the OLD FC6; the SrLogic
  runs BEFORE the two re-prefetches so a cleared S makes them run under the
  NEW user FC2 — the (6,6,6) vs (6,2,2) FC sequence is pinned to the data.
  RESET(0x4E70) -> [Internal(4), Internal(124), Prefetch] (len 132).
- All four files are supervisor-only in the vendored data. The mid-instruction
  FC switch IS gate-exercised (per-case transaction stream); the user-mode
  privilege-violation entry is correctness-only (not gated), documented in the
  runner. Both shapes pass through both drivers + a snapshot/restore anchor
  (the FC switch and the n124 idle round-trip at every micro-op boundary).
- Vendor ANDItoSR/ORItoSR/EORItoSR/RESET (8065 cases each, all in scope);
  runner FILES + sha256 + fetch-tests.sh extended; threshold raised
  157313 -> 189573 (the measured covered count).

Verified: determinism_gate + proptests, fmt --check, clippy -D warnings,
cargo test --workspace all green.

+++ crates/oracle-core/src/m68000/exception.rs
@@ -150,7 +150,9 @@ const CHK_SAVE_SR_SLOT: Slot = 1;
 /// width. Total 17 micro-ops (≤ `MAX_OPS`).
 pub(crate) fn build_chk_frame(buf: &mut RecipeBuf, idle: u8) {
     // The leading idle (n4 if Dn>bound, else n6) — the only per-trap parameter.
-    buf.push(MicroOp::Internal { cycles: idle });
+    buf.push(MicroOp::Internal {
+        cycles: idle as u16,
+    });
     // Capture the live SR (with CHK's N) + enter supervisor (set S, clear T).
     buf.push(MicroOp::EnterException {
         save_sr: CHK_SAVE_SR_SLOT,
